+++ src/gui/central_widget.cpp
@@ -24,16 +24,35 @@
  * @param mw pointer to MainWindow object
  */
 CentralWidget::CentralWidget(MainWindow* mw) {
-    QHBoxLayout *layout = new QHBoxLayout;
-    this->setLayout(layout);
+    this->main_layout = new QHBoxLayout;
+    this->main_layout->setSpacing(12);
+    this->setLayout(this->main_layout);
+
+    this->practice_diagram_panel = new QWidget(this);
+    QVBoxLayout *diagram_layout = new QVBoxLayout(this->practice_diagram_panel);
+    diagram_layout->setContentsMargins(0, 0, 0, 0);
+    QHBoxLayout *diagram_toolbar = new QHBoxLayout;
+    diagram_toolbar->addStretch();
+    QPushButton *full_tree_button = new QPushButton(tr("Full tree…"), this->practice_diagram_panel);
+    full_tree_button->setToolTip(tr("Open the complete point-group decision tree"));
+    diagram_toolbar->addWidget(full_tree_button);
+    diagram_layout->addLayout(diagram_toolbar);
+
+    this->practice_diagram = new PracticeFlowchartDiagram(this->practice_diagram_panel);
+    this->practice_diagram->setMinimumWidth(540);
+    diagram_layout->addWidget(this->practice_diagram);
+    this->practice_diagram_panel->setVisible(false);
+    this->main_layout->addWidget(this->practice_diagram_panel, 5);
+    connect(full_tree_button, &QPushButton::clicked, this, &CentralWidget::show_full_flowchart);
 
     this->gl_widget = new GLWidget(this);
     this->gl_widget->setSizePolicy(QSizePolicy::Policy::Expanding, QSizePolicy::Policy::Expanding);
-    layout->addWidget(this->gl_widget, 3);
+    this->main_layout->addWidget(this->gl_widget, 6);
 
     this->side_widget = new QStackedWidget(this);
     this->side_widget->setFrameShape(QFrame::NoFrame);
-    layout->addWidget(this->side_widget, 1);
+    this->side_widget->setMinimumWidth(310);
+    this->main_layout->addWidget(this->side_widget, 4);
 
     this->symmetry_widget = new SymmetryWidget(this);
     side_widget->addWidget(this->symmetry_widget);
@@ -52,6 +71,10 @@ CentralWidget::CentralWidget(MainWindow* mw) {
     connect(this->practice_widget, SIGNAL(open_character_table_dialog()), mw, SLOT(open_character_table_dialog()));
     connect(this->practice_widget, SIGNAL(highlight_atoms(QList<unsigned int>)), this->gl_widget, SLOT(highlight_atoms(QList<unsigned int>)));
     connect(this->practice_widget, SIGNAL(label_atoms(QMap<unsigned int, std::string>)), this->gl_widget, SLOT(label_atoms(QMap<unsigned int, std::string>)));
+    connect(this->practice_widget, &PracticeWidget::flowchart_route_changed,
+            this->practice_diagram, &PracticeFlowchartDiagram::set_path);
+    connect(this->practice_widget, &PracticeWidget::flowchart_visibility_changed,
+            this, &CentralWidget::set_flowchart_visible);
 }
 
 /**
@@ -78,9 +101,11 @@ void CentralWidget::set_gui_mode(QAction* action) {
     switch (this->gui_mode) {
         case GuiMode::SymmetryViewer:
             index = this->side_widget->indexOf(this->symmetry_widget);
+            this->set_flowchart_visible(false);
             break;
         case GuiMode::Practice:
             index = this->side_widget->indexOf(this->practice_widget);
+            this->set_flowchart_visible(this->practice_widget->get_flowchart_visible());
             break;
         default:
             throw std::runtime_error("Unexpected GUI mode encountered.");
@@ -123,4 +148,34 @@ std::shared_ptr<Symmetry>& CentralWidget::get_symmetry() {
  */
 std::shared_ptr<Library>& CentralWidget::get_library() {
     return this->library;
+}
+
+void CentralWidget::start_current_structure_flowchart() {
+    this->practice_widget->start_current_structure_flowchart();
+    this->side_widget->setCurrentWidget(this->practice_widget);
+}
+
+void CentralWidget::set_flowchart_visible(bool visible) {
+    this->practice_diagram_panel->setVisible(visible);
+    this->main_layout->setStretch(0, visible ? 6 : 0);
+    this->main_layout->setStretch(1, visible ? 6 : 3);
+    this->main_layout->setStretch(2, visible ? 4 : 1);
+}
+
+void CentralWidget::show_full_flowchart() {
+    QDialog *dialog = new QDialog(this);
+    dialog->setAttribute(Qt::WA_DeleteOnClose);
+    dialog->setWindowTitle(tr("Complete point-group decision tree"));
+    dialog->resize(1250, 820);
+
+    QVBoxLayout *layout = new QVBoxLayout(dialog);
+    PracticeFlowchartDiagram *overview = new PracticeFlowchartDiagram(dialog);
+    overview->set_overview_mode(true);
+    overview->set_path(this->practice_diagram->get_path());
+    layout->addWidget(overview);
+
+    QPushButton *close_button = new QPushButton(tr("Close"), dialog);
+    connect(close_button, &QPushButton::clicked, dialog, &QDialog::accept);
+    layout->addWidget(close_button, 0, Qt::AlignRight);
+    dialog->show();
 }